[core] Split fuzz tests into separate file
